split out v1 vs v3 execution code
